StudyMatch-backend/study_match_rag/requirements.txt
unstructured
tiktoken
pinecone-client
pypdf
langchain
langchain_community
pandas
numpy
python-dotenv
langchain_pinecone
langchain_ollama
langchain_groq